+++ doc/changelog.md
@@ -6,8 +6,9 @@ This files contains a version history including all changes relevant for semanti
 
 
 ## Upcoming Changes
+ - new Feature: added AutoValue Feature to repository implementations
+ - Fixed Bug in linq ExpressionTreeBuilder (constantValue)
 
-*(none)*